Add a button which generates new grid
Add a button which generates a new grid based on
user input each time it is pressed. The old grid
will be removed and replaced with a new grid
which will take up exactly the same amount of
space as the old one. Space inside of a grid will
be distributed equally among the number of squares
user decided to input.